Troubleshooting array issue
