+++ src/style.css
@@ -98,7 +98,3 @@ textarea {
 .view {
   width: 100%;
 }
-
-.fa-square {
-  cursor: pointer;
-}
